Duplication, reloading + effects
Made it so when you kill an enemy two projectiles fire in the air and an enemy spawns at each of them (AKA duplication which was the main idea). Also added gun reload functionality with the appropriate animation and sounds. Known issue - the enemy has too many polygons and when they duplicate it tanks performance :(